std::process fuchsia support cleanup
